refactor-#35: move activity to application
