+++ src/app.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { CqrsModule } from '@nestjs/cqrs';
 
 import { DatabaseModule } from './database/database.module';
 import { CatalogModule } from './catalog/catalog.module';
@@ -10,7 +9,6 @@ import { ElasticSearchModule } from './elastic-search/elastic-search.module';
 @Module({
   imports: [
     ConfigModule.forRoot(),
-    CqrsModule,
     ElasticSearchModule,
     DatabaseModule,
     CatalogModule,
