+++ src/catalog/application/application.module.ts
@@ -1,4 +1,20 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
+import { CqrsModule } from '@nestjs/cqrs';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-@Module({})
+import { ActivityRepository } from '../infrastructure/repositories';
+import { CommandHandlers } from './commands/handlers';
+import { ActivityController } from './controllers';
+import { QueryHandlers } from './queries/handlers';
+import { ActivityService } from './services';
+
+@Global()
+@Module({
+    imports: [
+        TypeOrmModule.forFeature([ActivityRepository]),
+        CqrsModule,
+    ],
+    controllers: [ActivityController],
+    providers: [ActivityService, ...CommandHandlers, ...QueryHandlers],
+})
 export class ApplicationModule {}
