+++ src/catalog/application/commands/handlers/index.ts
