+++ src/catalog/application/commands/handlers/update-activity.handler.ts
@@ -2,7 +2,7 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 import { RpcException } from '@nestjs/microservices';
 
-import { ElasticSearchService } from '../../../../elastic-search/services/';
+import { ElasticSearchService } from '../../../../elastic-search/services';
 import { ISearchBody } from '../../../../elastic-search/interfaces';
 import { ActivityRepository } from '../../../infrastructure/repositories';
 import { UpdateActivityCommand } from '../impl';
