+++ src/catalog/application/commands/impl/delete-activity.command.ts
