+++ src/catalog/application/commands/impl/update-activity.command.ts
