+++ src/catalog/application/dto/update-activity.dto.ts
