+++ src/catalog/application/models/text-response.model.ts
