+++ src/catalog/application/queries/handlers/get-activities-by-query.handler.ts
