+++ src/catalog/application/queries/handlers/get-all-activities.handler.ts
