+++ src/catalog/application/queries/handlers/get-single-activity.handler.ts
