+++ src/catalog/application/queries/handlers/index.ts
