+++ src/catalog/application/queries/impl/get-activities-by-query.query.ts
