+++ src/catalog/application/queries/impl/get-all-activities.query.ts
