+++ src/catalog/application/queries/impl/get-single-activity.query.ts
