+++ src/catalog/application/services/index.ts
@@ -1,0 +1,1 @@
+export * from './activity.service';
